Move tests for belongs_to :touch into belongs_to associations tests file
