+++ activerecord/test/cases/timestamp_test.rb
@@ -92,78 +92,6 @@ class TimestampTest < ActiveRecord::TestCase
     assert_nothing_raised { Car.first.touch }
   end
 
-  def test_saving_a_record_with_a_belongs_to_that_specifies_touching_the_parent_should_update_the_parent_updated_at
-    pet   = Pet.first
-    owner = pet.owner
-    previously_owner_updated_at = owner.updated_at
-
-    pet.name = "Fluffy the Third"
-    pet.save
-
-    assert_not_equal previously_owner_updated_at, pet.owner.updated_at
-  end
-
-  def test_destroying_a_record_with_a_belongs_to_that_specifies_touching_the_parent_should_update_the_parent_updated_at
-    pet   = Pet.first
-    owner = pet.owner
-    previously_owner_updated_at = owner.updated_at
-
-    pet.destroy
-
-    assert_not_equal previously_owner_updated_at, pet.owner.updated_at
-  end
-
-  def test_saving_a_record_with_a_belongs_to_that_specifies_touching_a_specific_attribute_the_parent_should_update_that_attribute
-    klass = Class.new(ActiveRecord::Base) do
-      def self.name; 'Pet'; end
-      belongs_to :owner, :touch => :happy_at
-    end
-
-    pet   = klass.first
-    owner = pet.owner
-    previously_owner_happy_at = owner.happy_at
-
-    pet.name = "Fluffy the Third"
-    pet.save
-
-    assert_not_equal previously_owner_happy_at, pet.owner.happy_at
-  end
-
-  def test_touching_a_record_with_a_belongs_to_that_uses_a_counter_cache_should_update_the_parent
-    klass = Class.new(ActiveRecord::Base) do
-      def self.name; 'Pet'; end
-      belongs_to :owner, :counter_cache => :use_count, :touch => true
-    end
-
-    pet = klass.first
-    owner = pet.owner
-    owner.update_columns(happy_at: 3.days.ago)
-    previously_owner_updated_at = owner.updated_at
-
-    pet.name = "I'm a parrot"
-    pet.save
-
-    assert_not_equal previously_owner_updated_at, pet.owner.updated_at
-  end
-
-  def test_touching_a_record_touches_parent_record_and_grandparent_record
-    klass = Class.new(ActiveRecord::Base) do
-      def self.name; 'Toy'; end
-      belongs_to :pet, :touch => true
-    end
-
-    toy = klass.first
-    pet = toy.pet
-    owner = pet.owner
-    time = 3.days.ago
-
-    owner.update_columns(updated_at: time)
-    toy.touch
-    owner.reload
-
-    assert_not_equal time, owner.updated_at
-  end
-
   def test_timestamp_attributes_for_create
     toy = Toy.first
     assert_equal toy.send(:timestamp_attributes_for_create), [:created_at, :created_on]
